Extract GetLastModified to Platform
