+++ src/openrct2/core/File.cpp
@@ -121,30 +121,7 @@ namespace File
 
     uint64_t GetLastModified(const std::string& path)
     {
-        uint64_t lastModified = 0;
-#ifdef _WIN32
-        auto pathW = String::ToWideChar(path);
-        auto hFile = CreateFileW(pathW.c_str(), GENERIC_READ, FILE_SHARE_READ, nullptr, OPEN_EXISTING, 0, nullptr);
-        if (hFile != INVALID_HANDLE_VALUE)
-        {
-            FILETIME ftCreate, ftAccess, ftWrite;
-            if (GetFileTime(hFile, &ftCreate, &ftAccess, &ftWrite))
-            {
-                lastModified = (static_cast<uint64_t>(ftWrite.dwHighDateTime) << 32ULL)
-                    | static_cast<uint64_t>(ftWrite.dwLowDateTime);
-            }
-            CloseHandle(hFile);
-        }
-#else
-        struct stat statInfo
-        {
-        };
-        if (stat(path.c_str(), &statInfo) == 0)
-        {
-            lastModified = statInfo.st_mtime;
-        }
-#endif
-        return lastModified;
+        return Platform::GetLastModified(path);
     }
 } // namespace File
