+++ src/openrct2/platform/Platform2.h
@@ -35,6 +35,7 @@ namespace Platform
     std::string GetCurrentExecutablePath();
     std::string GetCurrentExecutableDirectory();
     bool FileExists(const std::string path);
+    uint64_t GetLastModified(const std::string& path);
     rct2_time GetTimeLocal();
     rct2_date GetDateLocal();
     bool FindApp(const std::string& app, std::string* output);
